pyproject.toml
[project]
name = "switchbot-temperature-dashboard-for-pi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 88

exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle (error)
    "F",  # pyflakes
    "W",  # pycodestyle (warning)
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

ignore = [
    "E203", # black と競合
    "E501", # line-length は black/ruff-format に任せる
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
